Connect intervals of a first dataframe using a second dataframe of intervals

Connect the intervals of a first dataframe given that the can be considered connected if the separation between two of them are covered by a interval of a second dataframe.

Connect the intervals of a first dataframe given that the can be considered connected if the separation between two of them are covered by a interval of a second dataframe.

Usage

connectOverInterval(interval1, interval2)

connectOverInterval(interval1, interval2)

Arguments

interval1

first dataframe

interval2

second dataframe

Value

A dataframe of intervals representing the resulting connected intervals

A dataframe of intervals representing the resulting connected intervals